backend/requirement.txt
fastapi
uvicorn
openai
python-dotenv
pillow
reportlab